Update
check if already exists before cache

+++ index.js
@@ -1,17 +1,18 @@
 'use strict'
 
-const express = require('express')
-const request = require('request-promise')
-const cdnify = require('./upload')
-const low = require('lowdb')
-const storage = require('lowdb/file-sync')
 const fs = require('fs')
-const moment = require('moment')
 const path = require('path')
 
-const db = low('db.json', { storage })
-const app = express()
+const express = require('express')
+const request = require('request-promise')
+const low = require('lowdb')
+const storage = require('lowdb/file-sync')
+const moment = require('moment')
 
+const cdnify = require('./upload')
+
+const app = express()
+const db = low('db.json', { storage })
 const key = '3CZtMuShxwtbGP1bLL6S8DZZVBb7r6NkK8ltECfr'  // this is a personal key, please apply for one for yourself for free at https://api.nasa.gov
 const CDN_HOST = 'http://7xpbf9.com2.z0.glb.qiniucdn.com/' // this is a Chinese cloud storage provider, you should apply for your own too
 const PATH = path.join(__dirname, 'images')
@@ -61,41 +62,46 @@ const cache = (time) => {
     time = moment()
   }
 
-  const picName = `${moment().format('YYYY-MM-DD')}.jpg`
+  const picName = `${moment().format('YYYY-MM-DD')}_nasa_apod.jpg`
   const pathToSave = path.join(PATH, picName)
 
-  fetch(time)
-    .then(url => {
-      daysBefore = 0
+  let qiniuPicUrl = db('images').takeRight()[0].url.split('/');
+  let qiniuPicName = qiniuPicUrl[qiniuPicUrl.length - 1];
 
-      request(url)
-        .pipe(fs.createWriteStream(pathToSave))
-        .on('finish', () => {
-          cdnify(picName, pathToSave)
-            .then(reply => {
-              console.log('uploaded to qiniu, key: %s, hash: %s, at %s', reply.key, reply.hash, new Date())
-              latestImageLocation = CDN_HOST + reply.key
+  if (picName !== qiniuPicName) {
+    fetch(time)
+      .then(url => {
+        daysBefore = 0
 
-              db('images')
-                .chain()
-                .push({ url: CDN_HOST + reply.key })
-                .remove(element => {
-                  return latestImageLocation !== element.url
-                })
-                .value()
+        request(url)
+          .pipe(fs.createWriteStream(pathToSave))
+          .on('finish', () => {
+            cdnify(picName, pathToSave)
+              .then(reply => {
+                console.log('uploaded to qiniu, key: %s, hash: %s, at %s', reply.key, reply.hash, new Date())
+                latestImageLocation = CDN_HOST + reply.key
 
-              fs.unlinkSync(path.join(PATH, picName))
-            })
-            .catch(err => {
-              console.error(err)
-            })
-        })
-    })
-    .catch(err => {
-      console.error(err)
-      daysBefore++
-      cache(moment().subtract(daysBefore, 'days'))
-    })
+                db('images')
+                  .chain()
+                  .push({ url: CDN_HOST + reply.key })
+                  .remove(element => {
+                    return latestImageLocation !== element.url
+                  })
+                  .value()
+
+                fs.unlinkSync(path.join(PATH, picName))
+              })
+              .catch(err => {
+                console.error(err)
+              })
+          })
+      })
+      .catch(err => {
+        console.error(err)
+        daysBefore++
+        cache(moment().subtract(daysBefore, 'days'))
+      })
+  }
 }
 
 app.get('/apod', (req, res) => {
